A little bit of hook-refactoring
