+++ src/components/platforms/info/imageEditor/index.tsx
@@ -1,18 +1,23 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react"
+import { useCallback, useRef, useState } from "react"
 import { useRecoilValue, useSetRecoilState } from "recoil"
 import { ImageBinSrcSelectorFamily } from "../../../../recoil/selectors"
-
-import { open, OpenDialogOptions } from "@tauri-apps/api/dialog"
-import { readBinaryFile, BaseDirectory } from "@tauri-apps/api/fs"
-
 import { Modal } from "../../../modal"
-
-import "./index.css"
 import { invokeSaveFile } from "../../../../utils/invokes"
 import {
   fileSystemInvalidationAtom,
   pocketPathAtom,
 } from "../../../../recoil/atoms"
+import { ImageInfo } from "./types"
+import {
+  renderStamps,
+  useAddNewImageCallback,
+  useCanvasToDragStamps,
+  useMoveStampCallback,
+  useSetCurrentImageAsStamp,
+  useUpdateImageStampValuesCallback,
+} from "./hooks/stamps"
+
+import "./index.css"
 
 type ImageEditorProps = {
   onClose: () => void
@@ -36,135 +41,14 @@ export const ImageEditor = ({
   const [imageStamps, setImageStamps] = useState<ImageInfo[]>([])
   const canvasRef = useRef<HTMLCanvasElement>(null)
 
-  useEffect(() => {
-    ;(async () => {
-      if (imageStamps.length !== 0) return
-      const image = new Image()
-      image.src = currentImageSrc
-      await new Promise((resolve) => (image.onload = resolve))
+  useSetCurrentImageAsStamp(imageStamps, currentImageSrc, setImageStamps)
+  useCanvasToDragStamps(canvasRef, setImageStamps, selectedStampIndex)
+  renderStamps(canvasRef, width, height, imageStamps)
 
-      setImageStamps([
-        {
-          x: 0,
-          y: 0,
-          src: currentImageSrc,
-          scale: 1,
-          rotate: 0,
-          image,
-        },
-      ])
-    })()
-  }, [setImageStamps, currentImageSrc])
-
-  useEffect(() => {
-    if (!canvasRef.current) return
-    let mouseDown = false
-
-    canvasRef.current.onmousedown = () => (mouseDown = true)
-    canvasRef.current.onmouseup = () => (mouseDown = false)
-    canvasRef.current.onmouseout = () => (mouseDown = false)
-
-    canvasRef.current.onmousemove = (e) => {
-      if (!mouseDown) return
-
-      setImageStamps((stamps) => {
-        const stamp = stamps[selectedStampIndex]
-        const currentIndex = stamps.findIndex((i) => i === stamp)
-        const updated = {
-          ...stamp,
-          x: stamp.x + e.movementX,
-          y: stamp.y + e.movementY,
-        }
-
-        const newStamps = [...stamps]
-        newStamps.splice(currentIndex, 1, updated)
-        return newStamps
-      })
-    }
-  }, [selectedStampIndex, setImageStamps, canvasRef])
-
-  useEffect(() => {
-    if (!canvasRef.current) return
-    const context = canvasRef.current.getContext("2d")
-    if (!context) return
-
-    context.fillStyle = "white"
-    context.fillRect(0, 0, width, height)
-
-    const reversed = [...imageStamps].reverse()
-
-    context.globalCompositeOperation = "luminosity"
-
-    reversed.forEach((i) => {
-      context.save()
-      const { image, scale, x, y, rotate } = i
-
-      context.translate(width / 2 + x, height / 2 + y)
-      context.rotate(rotate * (Math.PI / 180))
-      context.translate(-(width / 2 + x), -(height / 2 + y))
-
-      context.drawImage(
-        i.image,
-        width / 2 + x - (image.width / 2) * scale,
-        height / 2 + y - (image.height / 2) * scale,
-        image.width * scale,
-        image.height * scale
-      )
-      context.restore()
-    })
-  }, [imageStamps])
-
-  const updateImageStampValues = useCallback(
-    (stamp: ImageInfo, updates: Partial<ImageInfo>) => {
-      setImageStamps((stamps) => {
-        const currentIndex = stamps.findIndex((i) => i === stamp)
-        const updated = { ...stamp, ...updates }
-
-        const newStamps = [...stamps]
-        newStamps.splice(currentIndex, 1, updated)
-        return newStamps
-      })
-    },
-    [setImageStamps]
-  )
-
-  const addNewImage = useCallback(async () => {
-    const options: OpenDialogOptions = {
-      multiple: false,
-      filters: [
-        {
-          name: "Image",
-          extensions: ["png", "jpeg", "jpg"],
-        },
-      ],
-    }
-    const imagePath = await open(options)
-
-    if (!imagePath || Array.isArray(imagePath)) return
-    const imageData = await readBinaryFile(imagePath)
-
-    const [name] = imagePath.split("/").reverse()
-    const file = new File([imageData], name)
-
-    const url = await URL.createObjectURL(file)
-    const image = new Image()
-    image.src = url
-    await new Promise((resolve) => (image.onload = resolve))
-
-    setImageStamps((stamps) => {
-      return [
-        {
-          x: 0,
-          y: 0,
-          src: url,
-          scale: 1,
-          rotate: 0,
-          image: image,
-        },
-        ...stamps,
-      ]
-    })
-  }, [])
+  const updateImageStampValues =
+    useUpdateImageStampValuesCallback(setImageStamps)
+  const addNewImage = useAddNewImageCallback(setImageStamps)
+  const moveStamp = useMoveStampCallback(setImageStamps, setSelectedStampIndex)
 
   const saveBinary = useCallback(async () => {
     if (!canvasRef.current) return
@@ -186,25 +70,9 @@ export const ImageEditor = ({
     }
 
     await invokeSaveFile(`${pocketPath}/${path}`, buffer)
-
     invalidateFS(Date.now())
-
     onClose()
   }, [width, height])
-
-  const moveStamp = useCallback(
-    (stamp: ImageInfo, move: -1 | 1) => {
-      setImageStamps((stamps) => {
-        const currentIndex = stamps.findIndex((s) => s === stamp)
-        const newStamps = stamps.filter((s) => s !== stamp)
-        newStamps.splice(currentIndex + move, 0, stamp)
-        return newStamps
-      })
-
-      setSelectedStampIndex((i) => i + move)
-    },
-    [setImageStamps, setSelectedStampIndex]
-  )
 
   return (
     <Modal className="image-editor">
@@ -273,7 +141,7 @@ export const ImageEditor = ({
                       )
                     }}
                   >
-                    Remove
+                    {"Remove"}
                   </button>
                   <button
                     onClick={() =>
@@ -285,7 +153,7 @@ export const ImageEditor = ({
                       })
                     }
                   >
-                    Reset
+                    {"Reset"}
                   </button>
                 </div>
                 <div className="image-editor__stamp-controls">
@@ -306,23 +174,14 @@ export const ImageEditor = ({
             )
           })}
         </div>
-        <button onClick={addNewImage}>Import Image</button>
+        <button onClick={addNewImage}>{"Import Image"}</button>
       </div>
       <div className="image-editor__bottom-buttons">
-        <button onClick={saveBinary}>Save</button>
+        <button onClick={saveBinary}>{"Save"}</button>
         <button className="image-editor__close-button" onClick={onClose}>
-          Cancel
+          {"Cancel"}
         </button>
       </div>
     </Modal>
   )
-}
-
-type ImageInfo = {
-  src: string
-  x: number
-  y: number
-  scale: number
-  rotate: number
-  image: HTMLImageElement
 }
