+++ src/components/platforms/info/imageEditor/types.ts
@@ -1,0 +1,8 @@
+export type ImageInfo = {
+  src: string
+  x: number
+  y: number
+  scale: number
+  rotate: number
+  image: HTMLImageElement
+}
